wip: submission entity

+++ internal/models/data/game.go
@@ -5,9 +5,9 @@ type Game struct {
 	GameId                 int64  `xorm:"'id' pk autoincr" json:"id"`
 	Title                  string `xorm:"'title' varchar(64) notnull" json:"title"`                                      // 比赛标题
 	Description            string `xorm:"'description' text" json:"description"`                                         // 比赛描述
-	MemberLimitMin         int    `xorm:"'member_limit_min' notnull default(1)" json:"member_limit_min"`                 // 比赛人数最小值
-	MemberLimitMax         int    `xorm:"'member_limit_max' default(10)" json:"member_limit_max"`                        // 比赛人数最大值
-	ParallelContainerLimit int    `xorm:"'parallel_container_limit' notnull default(2)" json:"parallel_container_limit"` // 容器并行限制
+	MemberLimitMin         int64  `xorm:"'member_limit_min' notnull default(1)" json:"member_limit_min"`                 // 比赛人数最小值
+	MemberLimitMax         int64  `xorm:"'member_limit_max' default(10)" json:"member_limit_max"`                        // 比赛人数最大值
+	ParallelContainerLimit int64  `xorm:"'parallel_container_limit' notnull default(2)" json:"parallel_container_limit"` // 容器并行限制
 	IsNeedWriteUp          bool   `xorm:"'is_need_write_up' bool default(1) notnull" json:"is_need_write_up"`            // 是否需要提交 WP
 	StartedAt              int64  `xorm:"'started_at' int notnull" json:"started_at"`                                    // 比赛开始时间
 	EndedAt                int64  `xorm:"'ended_at' int notnull" json:"ended_at"`                                        // 比赛结束时间
